Consensus: Cleanup: Destroy pow.o and move GetBlockProof(CBlockIndex) to chain.o
