+++ src/Makefile.am
@@ -107,7 +107,6 @@ BITCOIN_CORE_H = \
   netbase.h \
   net.h \
   noui.h \
-  pow.h \
   primitives/block.h \
   primitives/transaction.h \
   protocol.h \
@@ -184,7 +183,6 @@ libbitcoin_server_a_SOURCES = \
   miner.cpp \
   net.cpp \
   noui.cpp \
-  pow.cpp \
   rest.cpp \
   rpcblockchain.cpp \
   rpcmining.cpp \
